rename core.ts in skillmap, fix spacing
